% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_rohde.R
\name{adder}
\alias{adder}
\title{Add two functions}
\usage{
adder(a, b)
}
\arguments{
\item{a}{number}

\item{b}{number}
}
\value{
number
}
\description{
add together these two numbers
}
\examples{
adder(2,7)
}
