% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_note.R
\name{describe_note}
\alias{describe_note}
\title{Return \code{Hmisc::describe} in a Quarto callout note}
\usage{
describe_note(data, label = "Expand for an overview of data values")
}
\arguments{
\item{data}{A data frame as input to \code{Hmisc::describe()}.}

\item{label}{A label for the note heading}
}
\description{
Note: You must set \verb{#| output = asis} within the Quarto block
}
\examples{
describe_note(mtcars)
describe_note(mtcars, label = "Data values in the `mtcars` dataset")
}
