% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_rohde.R
\name{theme_rohde}
\alias{theme_rohde}
\title{Personal plotting function}
\usage{
theme_rohde(font_size = 12, font_family = "Lato")
}
\arguments{
\item{font_family}{}
}
\value{
ggplot2 plot
}
\description{
Personal plotting function
}
